+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.0 (unreleased)
 
 - Updated Polars to 0.29.0
+- Added support for `List` and `Struct` to `to_a` method
 - Added support for creating series from Numo arrays
 - Added column assignment to `DataFrame`
 - Added `sort!` and `to_a` methods to `DataFrame`